requirements.txt
streamlit
requests
python-dotenv
pydantic
tenacity
google-genai
